update Wikimedia rule to reflect native HTTPS support
This patch comes from upstream HTTPS-Everywhere extension. According
to its git log, authors are:
 - Roan Kattouw
 - Jérémie Roquet

From the file:

Wikipedia and other Wikimedia Foundation wikis previously had no real
HTTPS support, and URLs had to be rewritten to
https://secure.wikimedia.org/$wikitype/$language/

This is no longer the case, see
https://blog.wikimedia.org/2011/10/03/native-https-support-enabled-for-all-wikim
so this file is a lot simpler these days.